Extract the base Intercepting logic into a base class, OutputTypeValueProviderBase
This base class will now be inherited by OutputType and OutputTypeEx value providers for BrowseObject
