+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Properties/InterceptedProjectProperties/OutputTypeExValueProvider.cs
@@ -2,57 +2,44 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.ComponentModel.Composition;
-using System.Threading.Tasks;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.Properties
 {
     /// OutputTypeEx acts as a converter for the MSBuild OutputType value expressed as <see cref="VSLangProj110.prjOutputTypeEx"/>.
     [ExportInterceptingPropertyValueProvider("OutputTypeEx", ExportInterceptingPropertyValueProviderFile.ProjectFile)]
-    internal sealed class OutputTypeExValueProvider : InterceptingPropertyValueProviderBase
+    internal sealed class OutputTypeExValueProvider : OutputTypeValueProviderBase
     {
         private readonly ProjectProperties _properties;
 
-        private readonly Dictionary<string, string> _getOutputTypeExMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+        private static readonly ImmutableDictionary<string, string> s_getOutputTypeExMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
         {
             {"WinExe",          "0" },
             {"Exe",             "1" },
             {"Library",         "2" },
             {"AppContainerExe", "3" },
             {"WinMDObj",        "4" },
-        };
+        }.ToImmutableDictionary();
 
-        private readonly Dictionary<string, string> _setOutputTypeExMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+        private static readonly ImmutableDictionary<string, string> s_setOutputTypeExMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
         {
             {"0", "WinExe" },
             {"1", "Exe" },
             {"2", "Library" },
             {"3", "AppContainerExe" },
             {"4", "WinMDObj"},
-        };
+        }.ToImmutableDictionary();
+
+        protected override ImmutableDictionary<string, string> GetMap => s_getOutputTypeExMap;
+
+        protected override ImmutableDictionary<string, string> SetMap => s_setOutputTypeExMap;
 
         [ImportingConstructor]
         public OutputTypeExValueProvider(ProjectProperties properties)
+            : base(properties)
         {
-            _properties = properties;
-        }
-
-        public override async Task<string> OnGetEvaluatedPropertyValueAsync(string evaluatedPropertyValue, IProjectProperties defaultProperties)
-        {
-            var configuration = await _properties.GetConfigurationGeneralPropertiesAsync().ConfigureAwait(false);
-            var value = await configuration.OutputType.GetEvaluatedValueAtEndAsync().ConfigureAwait(false);
-            return _getOutputTypeExMap[value];
-        }
-
-        public override async Task<string> OnSetPropertyValueAsync(string unevaluatedPropertyValue, IProjectProperties defaultProperties, IReadOnlyDictionary<string, string> dimensionalConditions = null)
-        {
-            var value = _setOutputTypeExMap[unevaluatedPropertyValue];
-            var configuration = await _properties.GetConfigurationGeneralPropertiesAsync().ConfigureAwait(false);
-            await configuration.OutputType.SetValueAsync(value).ConfigureAwait(false);
-
-            // We need to return null so we dont persist a value for the MSBuild property 'OutputTypeEx'
-            return null;
         }
     }
 }
